GUI supports most functions except score and changing speed
so i did change some of the car class but for the most part i kept the GUI components in their own classes that i can eventually make extend your work. trackAnimate can extend the track class and draw car extends car.  how do we want to handle the different speeds of the cars? i was thinking we could have different car types that the user can select along with the color selection and make each car type a different visual. the car types could then have different probabilities of getting random speed values.... another possability, though id have to change the animation for this would be to have it randomly select a speed at each point and go that speed till it reaches the next point. this will be much harder to animate and gives the user less options but is more on par with assignment discription. what do you guys think?